Created component folder and Moved components to their own files
